requirements/base.txt
Django==1.7.5
PyYAML==3.11
South==0.7.3
